requirements.txt
python-telegram-bot
python-dotenv
google-genai
pillow
pandas
openpyxl
gspread
oauth2client